+++ docs/.vuepress/client.ts
@@ -1,7 +1,6 @@
 import { defineGiscusConfig } from '@vuepress/plugin-comment/client'
 import { defineDocSearchConfig } from '@vuepress/plugin-docsearch/client'
 import { defineClientConfig } from 'vuepress/client'
-import type { ClientConfig } from 'vuepress/client'
 import CommentPage from './layouts/CommentPage.vue'
 
 defineGiscusConfig({
@@ -66,4 +65,4 @@ export default defineClientConfig({
     // We override the default layout to provide comment service
     CommentPage,
   },
-}) as ClientConfig
+})