+++ plugins/development/plugin-active-header-links/src/client/config.ts
@@ -1,5 +1,4 @@
 import { defineClientConfig } from 'vuepress/client'
-import type { ClientConfig } from 'vuepress/client'
 import { useActiveHeaderLinks } from './composables/index.js'
 
 declare const __AHL_HEADER_LINK_SELECTOR__: string
@@ -23,4 +22,4 @@ export default defineClientConfig({
       offset,
     })
   },
-}) as ClientConfig
+})