+++ plugins/features/plugin-notice/src/client/config.ts
@@ -1,6 +1,5 @@
 import { h } from 'vue'
 import { defineClientConfig } from 'vuepress/client'
-import type { ClientConfig } from 'vuepress/client'
 import type { NoticeAttrOptions } from '../shared/index.js'
 import { Notice } from './components/index.js'
 
@@ -10,4 +9,4 @@ declare const __NOTICE_OPTIONS__: NoticeAttrOptions[]
 
 export default defineClientConfig({
   rootComponents: [() => h(Notice, { config: __NOTICE_OPTIONS__ })],
-}) as ClientConfig
+})