+++ plugins/markdown/plugin-prismjs/src/node/prismjsPlugin.ts
@@ -1,28 +1,11 @@
 import { lineNumbers as lineNumbersPlugin } from '@vuepress/highlighter-helper'
-import type { MarkdownItLineNumbersOptions } from '@vuepress/highlighter-helper'
 import type { Plugin } from 'vuepress/core'
 import { loadLanguages } from './loadLanguages.js'
 import { highlightPlugin, preWrapperPlugin } from './markdown/index.js'
+import type { PrismjsPluginOptions } from './options.js'
+import { prepareConfigFile } from './prepareConfigFile.js'
 import { resolveHighlighter } from './resolveHighlighter.js'
 import type { HighlightOptions, PreWrapperOptions } from './types.js'
-
-/**
- * Options of @vuepress/plugin-prismjs
- */
-export interface PrismjsPluginOptions
-  extends Pick<MarkdownItLineNumbersOptions, 'lineNumbers'>,
-    PreWrapperOptions,
-    HighlightOptions {
-  /**
-   * Languages to preload
-   *
-   * Workaround for prismjs language reloading issue
-   *
-   * @default ['markdown', 'jsdoc', 'yaml']
-   * @see https://github.com/PrismJS/prism/issues/2716
-   */
-  preloadLanguages?: string[]
-}
 
 export const prismjsPlugin = ({
   preloadLanguages = ['markdown', 'jsdoc', 'yaml'],
@@ -48,4 +31,6 @@ export const prismjsPlugin = ({
       md.use(lineNumbersPlugin, { lineNumbers, removeLastLine: true })
     }
   },
+
+  clientConfigFile: (app) => prepareConfigFile(app, options),
 })