+++ plugins/markdown/plugin-shiki/src/node/transformers/getTransformers.ts
@@ -1,9 +1,14 @@
 import {
+  transformerMetaWordHighlight,
   transformerNotationDiff,
   transformerNotationErrorLevel,
   transformerNotationFocus,
   transformerNotationHighlight,
+  transformerNotationWordHighlight,
+  transformerRenderWhitespace,
 } from '@shikijs/transformers'
+import type { WhitespacePosition } from '@vuepress/highlighter-helper'
+import { resolveWhitespacePosition } from '@vuepress/highlighter-helper'
 import type { ShikiTransformer } from 'shiki'
 import type { ShikiHighlightOptions } from '../types.js'
 import {
@@ -39,6 +44,11 @@ export const getTransformers = (
     transformers.push(transformerNotationErrorLevel())
   }
 
+  if (options.notationWordHighlight) {
+    transformers.push(transformerNotationWordHighlight())
+    transformers.push(transformerMetaWordHighlight())
+  }
+
   transformers.push(
     addClassTransformer,
     cleanupTransformer,
@@ -47,4 +57,15 @@ export const getTransformers = (
   )
 
   return transformers
+}
+
+export const whitespaceTransformer = (
+  meta: string,
+  defaultPosition: boolean | WhitespacePosition = false,
+): ShikiTransformer[] => {
+  const position = resolveWhitespacePosition(meta, defaultPosition)
+  // disabled current code block
+  if (position === false) return []
+
+  return [transformerRenderWhitespace({ position })]
 }