+++ themes/theme-default/src/client/config.ts
@@ -4,7 +4,6 @@ import { setupSidebarItems } from '@theme/useSidebarItems'
 import { hasGlobalComponent } from '@vuepress/helper/client'
 import { h } from 'vue'
 import { defineClientConfig } from 'vuepress/client'
-import type { ClientConfig } from 'vuepress/client'
 import { Badge, CodeGroup, CodeGroupItem } from './components/global/index.js'
 import Layout from './layouts/Layout.vue'
 import NotFound from './layouts/NotFound.vue'
@@ -46,4 +45,4 @@ export default defineClientConfig({
     Layout,
     NotFound,
   },
-}) as ClientConfig
+})