+++ themes/theme-default/src/node/defaultTheme.ts
@@ -1,3 +1,8 @@
+import {
+  addViteConfig,
+  addViteOptimizeDepsExclude,
+  chainWebpack,
+} from '@vuepress/helper'
 import { activeHeaderLinksPlugin } from '@vuepress/plugin-active-header-links'
 import { backToTopPlugin } from '@vuepress/plugin-back-to-top'
 import { copyCodePlugin } from '@vuepress/plugin-copy-code'
@@ -75,6 +80,52 @@ export const defaultTheme = ({
     },
 
     clientConfigFile: path.resolve(__dirname, '../client/config.js'),
+
+    extendsBundlerOptions: (bundlerOptions, app) => {
+      // FIXME: hide sass deprecation warning for mixed-decls
+      addViteConfig(bundlerOptions, app, {
+        css: {
+          preprocessorOptions: {
+            sass: {
+              logger: {
+                warn: (message, { deprecation, deprecationType }) => {
+                  if (deprecation && deprecationType.id === 'mixed-decls')
+                    return
+
+                  console.warn(message)
+                },
+              },
+            },
+            scss: {
+              logger: {
+                warn: (message, { deprecation, deprecationType }) => {
+                  if (deprecation && deprecationType.id === 'mixed-decls')
+                    return
+
+                  console.warn(message)
+                },
+              },
+            },
+          },
+        },
+      })
+      chainWebpack(bundlerOptions, app, (config) => {
+        config.module
+          .rule('scss')
+          .use('sass-loader')
+          .tap((options) => ({
+            api: 'modern-compiler',
+            ...options,
+            sassOptions: {
+              silenceDeprecations: ['mixed-decls'],
+              ...options.sassOptions,
+            },
+          }))
+      })
+
+      // ensure theme alias is not optimized by Vite
+      addViteOptimizeDepsExclude(bundlerOptions, app, '@theme')
+    },
 
     extendsPage: (page: Page<Partial<DefaultThemePageData>>) => {
       // save relative file path into page data to generate edit link