+++ tools/highlighter-helper/src/node/index.ts
@@ -1,1 +1,2 @@
 export * from './lineNumbers/index.js'
+export * from './whitespace.js'